+++ scripts/redact_cli_py/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support to only redact specific labels.
+- Add support for one page pdfs in batch redact CLI (`batch_redact.py`)
 
 ## [0.2.1] - 2021-08-05
 ### Added